requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
matplotlib==3.7.2
